Extract `controlledSvgBezierCurveUtils.kt`
